Update
Kinect Gestures, Manager and Wrapper are used to detect gestures and
display it in unity. Listener is not necessary, but it could be used to
send commands to server.

VideoFeedUrl is a script for url field in unity.
Ffserver and Ffserver2 are similar, but F2 uses VideoFeedUrl as an
input, F doesn't.